% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uCAREChemSuiteCLI.R
\name{drug.class.deterministic}
\alias{drug.class.deterministic}
\title{drug.class.deterministic}
\usage{
drug.class.deterministic("sdf")
}
\arguments{
\item{sdf}{input sdf file}
}
\value{
predicted drug class of the candidate drug by deterministic model
}
\description{
Takes structure data file (SDF) of candidate drug and predicts its drug class using deterministic model.
}
\details{
uCAREChemSuiteCLI
}
\examples{
{
example.class.deterministic<- system.file('extdata/example.sdf', package="uCAREChemSuiteCLI")
drug.class.deterministic(example.class.deterministic)
}
}
