% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uCAREChemSuiteCLI.R
\name{drug.class.stochastic}
\alias{drug.class.stochastic}
\title{drug.class.stochastic}
\usage{
drug.class.stochastic("sdf", "NearestNeighbor", "Threshold")
}
\arguments{
\item{sdf}{input sdf file}

\item{NearestNeighbor}{Nearest Neighbor = 1, 3}

\item{Threshold}{Threshold = 0.25, 0.3, 0.35, 0.4}
}
\value{
Predicted drug class of the candidate drug using Nearest Neighbor algorithm
}
\description{
Takes structure data file (SDF) of candidate drug, Nearest Neighbor value and threshold similarity score to predict its drug class using stochastic model.
}
\details{
uCAREChemSuiteCLI
}
\examples{
{
example.class.stochastic<- system.file('extdata/example.sdf', package="uCAREChemSuiteCLI")
drug.class.stochastic(example.class.stochastic,"3","0.25")
}
}
