% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uCAREChemSuiteCLI.R
\name{drug.resistome.deterministic}
\alias{drug.resistome.deterministic}
\title{drug.resistome.deterministic}
\usage{
drug.resistome.deterministic("sdf")
}
\arguments{
\item{sdf}{input sdf file}
}
\value{
Predicted resistome of the candidate drug using deterministic model
}
\description{
Takes structure data file (SDF) of candidate drug to predicts its resistome using deterministic model.
}
\details{
uCAREChemSuiteCLI
}
\examples{
{
example.resistome.deterministic<- system.file('extdata/example.sdf', package="uCAREChemSuiteCLI")
drug.resistome.deterministic(example.resistome.deterministic)
}
}
